+++ examples/fling/main.go
@@ -1,8 +1,6 @@
 package main
 
 import (
-	"os"
-
 	"go.bbkane.com/warg"
 	"go.bbkane.com/warg/command"
 	"go.bbkane.com/warg/flag"
@@ -69,15 +67,15 @@ func app() *warg.App {
 				unlink,
 				command.ExistingFlags(linkUnlinkFlags),
 			),
+			section.ExistingCommand("version", warg.VersionCommand()),
 			section.Footer("Homepage: https://github.com/bbkane/fling"),
+			section.ExistingFlag("--color", warg.ColorFlag()),
 		),
-		warg.AddColorFlag(),
-		warg.AddVersionCommand(version),
 		warg.SkipValidation(),
 	)
 	return &app
 }
 
 func main() {
-	app().MustRun(os.Args, os.LookupEnv)
+	app().MustRun()
 }